start_code/instagram_automation/requirements.txt
instagrapi==2.2.1

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Task Queue
apscheduler==3.10.4
redis==5.0.1

# AI Services
openai==1.3.0
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Proxy support
httpx[socks]==0.25.2
